// packages/core/src/amazonqFeatureDev/client/featuredevproxyclient.d.ts
  export interface ChatAddMessageEvent {
    conversationId: ConversationId;
    messageId: MessageId;
    customizationArn?: CustomizationArn;
    userIntent?: UserIntent;
    hasCodeSnippet?: Boolean;
    programmingLanguage?: ProgrammingLanguage;
    activeEditorTotalCharacters?: Integer;
    timeToFirstChunkMilliseconds?: Double;
    timeBetweenChunks?: timeBetweenChunks;
    fullResponselatency?: Double;
    requestLength?: Integer;
    responseLength?: Integer;
    numberOfCodeBlocks?: Integer;
    hasProjectLevelContext?: Boolean;
  }
  export type ChatHistory = ChatMessage[];
  export interface ChatInteractWithMessageEvent {
    conversationId: ConversationId;
    messageId: MessageId;
    customizationArn?: CustomizationArn;
    interactionType?: ChatMessageInteractionType;
    interactionTarget?: ChatInteractWithMessageEventInteractionTargetString;
    acceptedCharacterCount?: Integer;
    acceptedLineCount?: Integer;
    acceptedSnippetHasReference?: Boolean;
    hasProjectLevelContext?: Boolean;
    userIntent?: UserIntent;
    addedIdeDiagnostics?: IdeDiagnosticList;
    removedIdeDiagnostics?: IdeDiagnosticList;
  }
// ...
  export interface ChatUserModificationEvent {
    conversationId: ConversationId;
    customizationArn?: CustomizationArn;
    messageId: MessageId;
    programmingLanguage?: ProgrammingLanguage;
    modificationPercentage: Double;
    hasProjectLevelContext?: Boolean;
  }
  export type ClientId = string;
  export type CodeAnalysisFindingsSchema = "codeanalysis/findings/1.0"|string;
  export type CodeAnalysisScope = "FILE"|"PROJECT"|string;
  export type CodeAnalysisStatus = "Completed"|"Pending"|"Failed"|string;
  export interface CodeAnalysisUploadContext {
    codeScanName: CodeScanName;
  }
  export interface CodeCoverageEvent {
    customizationArn?: CustomizationArn;
    programmingLanguage: ProgrammingLanguage;
    acceptedCharacterCount: PrimitiveInteger;
    totalCharacterCount: PrimitiveInteger;
    timestamp: Timestamp;
    unmodifiedAcceptedCharacterCount?: PrimitiveInteger;
    totalNewCodeCharacterCount?: PrimitiveInteger;
    totalNewCodeLineCount?: PrimitiveInteger;
    userWrittenCodeCharacterCount?: CodeCoverageEventUserWrittenCodeCharacterCountInteger;
    userWrittenCodeLineCount?: CodeCoverageEventUserWrittenCodeLineCountInteger;
  }
  export type CodeCoverageEventUserWrittenCodeCharacterCountInteger = number;
  export type CodeCoverageEventUserWrittenCodeLineCountInteger = number;
  export interface CodeDescription {
    href: CodeDescriptionHrefString;
  }
  export type CodeDescriptionHrefString = string;
  export interface CodeFixAcceptanceEvent {
    jobId: String;
    ruleId?: String;
    detectorId?: String;
    findingId?: String;
    programmingLanguage?: ProgrammingLanguage;
    linesOfCodeAccepted?: Integer;
    charsOfCodeAccepted?: Integer;
  }
  export interface CodeFixGenerationEvent {
    jobId: String;
    ruleId?: String;
    detectorId?: String;
    findingId?: String;
    programmingLanguage?: ProgrammingLanguage;
    linesOfCodeGenerated?: Integer;
    charsOfCodeGenerated?: Integer;
  }
// ...
  export interface CodeScanEvent {
    programmingLanguage: ProgrammingLanguage;
    codeScanJobId: CodeScanJobId;
    timestamp: Timestamp;
    codeAnalysisScope?: CodeAnalysisScope;
  }
  export interface CodeScanFailedEvent {
    programmingLanguage: ProgrammingLanguage;
    codeScanJobId: CodeScanJobId;
    timestamp: Timestamp;
    codeAnalysisScope?: CodeAnalysisScope;
  }
  export type CodeScanJobId = string;
  export type CodeScanName = string;
  export interface CodeScanRemediationsEvent {
    programmingLanguage?: ProgrammingLanguage;
    CodeScanRemediationsEventType?: CodeScanRemediationsEventType;
    timestamp?: Timestamp;
    detectorId?: String;
    findingId?: String;
    ruleId?: String;
    component?: String;
    reason?: String;
    result?: String;
    includesFix?: Boolean;
  }
  export type CodeScanRemediationsEventType = "CODESCAN_ISSUE_HOVER"|"CODESCAN_ISSUE_APPLY_FIX"|"CODESCAN_ISSUE_VIEW_DETAILS"|string;
  export interface CodeScanSucceededEvent {
    programmingLanguage: ProgrammingLanguage;
    codeScanJobId: CodeScanJobId;
    timestamp: Timestamp;
    numberOfFindings: PrimitiveInteger;
    codeAnalysisScope?: CodeAnalysisScope;
  }
// ...
  export interface DocGenerationEvent {
    conversationId: ConversationId;
    numberOfAddChars?: PrimitiveInteger;
    numberOfAddLines?: PrimitiveInteger;
    numberOfAddFiles?: PrimitiveInteger;
    userDecision?: DocUserDecision;
    interactionType?: DocInteractionType;
    userIdentity?: String;
    numberOfNavigation?: PrimitiveInteger;
    folderLevel?: DocFolderLevel;
  }
  export type DocInteractionType = "GENERATE_README"|"UPDATE_README"|"EDIT_README"|string;
  export type DocUserDecision = "ACCEPT"|"REJECT"|string;
  export interface DocV2AcceptanceEvent {
    conversationId: ConversationId;
    numberOfAddedChars: DocV2AcceptanceEventNumberOfAddedCharsInteger;
    numberOfAddedLines: DocV2AcceptanceEventNumberOfAddedLinesInteger;
    numberOfAddedFiles: DocV2AcceptanceEventNumberOfAddedFilesInteger;
    userDecision: DocUserDecision;
    interactionType: DocInteractionType;
    numberOfNavigations: DocV2AcceptanceEventNumberOfNavigationsInteger;
    folderLevel: DocFolderLevel;
  }
  export type DocV2AcceptanceEventNumberOfAddedCharsInteger = number;
  export type DocV2AcceptanceEventNumberOfAddedFilesInteger = number;
  export type DocV2AcceptanceEventNumberOfAddedLinesInteger = number;
  export type DocV2AcceptanceEventNumberOfNavigationsInteger = number;
  export interface DocV2GenerationEvent {
    conversationId: ConversationId;
    numberOfGeneratedChars: DocV2GenerationEventNumberOfGeneratedCharsInteger;
    numberOfGeneratedLines: DocV2GenerationEventNumberOfGeneratedLinesInteger;
    numberOfGeneratedFiles: DocV2GenerationEventNumberOfGeneratedFilesInteger;
    interactionType?: DocInteractionType;
    numberOfNavigations?: DocV2GenerationEventNumberOfNavigationsInteger;
    folderLevel?: DocFolderLevel;
  }
// ...
  export interface FeatureDevCodeAcceptanceEvent {
    conversationId: ConversationId;
    linesOfCodeAccepted: FeatureDevCodeAcceptanceEventLinesOfCodeAcceptedInteger;
    charactersOfCodeAccepted: FeatureDevCodeAcceptanceEventCharactersOfCodeAcceptedInteger;
    programmingLanguage?: ProgrammingLanguage;
  }
  export type FeatureDevCodeAcceptanceEventCharactersOfCodeAcceptedInteger = number;
  export type FeatureDevCodeAcceptanceEventLinesOfCodeAcceptedInteger = number;
  export interface FeatureDevCodeGenerationEvent {
    conversationId: ConversationId;
    linesOfCodeGenerated: FeatureDevCodeGenerationEventLinesOfCodeGeneratedInteger;
    charactersOfCodeGenerated: FeatureDevCodeGenerationEventCharactersOfCodeGeneratedInteger;
    programmingLanguage?: ProgrammingLanguage;
  }
  export type FeatureDevCodeGenerationEventCharactersOfCodeGeneratedInteger = number;
  export type FeatureDevCodeGenerationEventLinesOfCodeGeneratedInteger = number;
  export interface FeatureDevEvent {
    conversationId: ConversationId;
  }
// ...
  export type IdeCategory = "JETBRAINS"|"VSCODE"|"CLI"|"JUPYTER_MD"|"JUPYTER_SM"|"ECLIPSE"|"VISUAL_STUDIO"|string;
  export interface IdeDiagnostic {
    range?: Range;
    source?: IdeDiagnosticSourceString;
    severity?: DiagnosticSeverity;
    ideDiagnosticType: IdeDiagnosticType;
  }
  export type IdeDiagnosticList = IdeDiagnostic[];
  export type IdeDiagnosticSourceString = string;
  export type IdeDiagnosticType = "SYNTAX_ERROR"|"TYPE_ERROR"|"REFERENCE_ERROR"|"BEST_PRACTICE"|"SECURITY"|"OTHER"|string;
  export type IdempotencyToken = string;
// ...
  export interface InlineChatEvent {
    requestId: UUID;
    timestamp: Timestamp;
    inputLength?: PrimitiveInteger;
    numSelectedLines?: PrimitiveInteger;
    numSuggestionAddChars?: PrimitiveInteger;
    numSuggestionAddLines?: PrimitiveInteger;
    numSuggestionDelChars?: PrimitiveInteger;
    numSuggestionDelLines?: PrimitiveInteger;
    codeIntent?: Boolean;
    userDecision?: InlineChatUserDecision;
    responseStartLatency?: Double;
    responseEndLatency?: Double;
    programmingLanguage?: ProgrammingLanguage;
  }
// ...
  export interface MetricData {
    metricName: MetricDataMetricNameString;
    metricValue: Double;
    timestamp: Timestamp;
    product: MetricDataProductString;
    dimensions?: DimensionList;
  }
  export type MetricDataMetricNameString = string;
  export type MetricDataProductString = string;
  export type NextToken = string;
  export type Notifications = NotificationsFeature[];
  export interface NotificationsFeature {
    feature: FeatureName;
    toggle: OptInFeatureToggle;
  }
  export type OperatingSystem = "MAC"|"WINDOWS"|"LINUX"|string;
  export type OptInFeatureToggle = "ON"|"OFF"|string;
  export interface OptInFeatures {
    promptLogging?: PromptLogging;
    byUserAnalytics?: ByUserAnalytics;
    dashboardAnalytics?: DashboardAnalytics;
    notifications?: Notifications;
    workspaceContext?: WorkspaceContext;
  }
  export type OptOutPreference = "OPTIN"|"OPTOUT"|string;
// ...
  export type ProfileArn = string;
// ...
  export interface SendTelemetryEventRequest {
    clientToken?: IdempotencyToken;
    telemetryEvent: TelemetryEvent;
    optOutPreference?: OptOutPreference;
    userContext?: UserContext;
    profileArn?: ProfileArn;
  }
// ...
  export type String = string;
// ...
  export interface TelemetryEvent {
    userTriggerDecisionEvent?: UserTriggerDecisionEvent;
    codeCoverageEvent?: CodeCoverageEvent;
    userModificationEvent?: UserModificationEvent;
    codeScanEvent?: CodeScanEvent;
    codeScanSucceededEvent?: CodeScanSucceededEvent;
    codeScanFailedEvent?: CodeScanFailedEvent;
    codeScanRemediationsEvent?: CodeScanRemediationsEvent;
    codeFixGenerationEvent?: CodeFixGenerationEvent;
    codeFixAcceptanceEvent?: CodeFixAcceptanceEvent;
    metricData?: MetricData;
    chatAddMessageEvent?: ChatAddMessageEvent;
    chatInteractWithMessageEvent?: ChatInteractWithMessageEvent;
    chatUserModificationEvent?: ChatUserModificationEvent;
    terminalUserInteractionEvent?: TerminalUserInteractionEvent;
    featureDevEvent?: FeatureDevEvent;
    featureDevCodeGenerationEvent?: FeatureDevCodeGenerationEvent;
    featureDevCodeAcceptanceEvent?: FeatureDevCodeAcceptanceEvent;
    inlineChatEvent?: InlineChatEvent;
    transformEvent?: TransformEvent;
    docGenerationEvent?: DocGenerationEvent;
    docV2GenerationEvent?: DocV2GenerationEvent;
    docV2AcceptanceEvent?: DocV2AcceptanceEvent;
    testGenerationEvent?: TestGenerationEvent;
  }
  export type TenantId = string;
  export interface TerminalUserInteractionEvent {
    terminalUserInteractionEventType?: TerminalUserInteractionEventType;
    terminal?: String;
    terminalVersion?: String;
    shell?: String;
    shellVersion?: String;
    duration?: Integer;
    timeToSuggestion?: Integer;
    isCompletionAccepted?: Boolean;
    cliToolCommand?: String;
  }
  export type TerminalUserInteractionEventType = "CODEWHISPERER_TERMINAL_TRANSLATION_ACTION"|"CODEWHISPERER_TERMINAL_COMPLETION_INSERTED"|string;
  export interface TestGenerationEvent {
    jobId: UUID;
    groupName: TestGenerationJobGroupName;
    timestamp?: Timestamp;
    ideCategory?: IdeCategory;
    programmingLanguage?: ProgrammingLanguage;
    numberOfUnitTestCasesGenerated?: Integer;
    numberOfUnitTestCasesAccepted?: Integer;
    linesOfCodeGenerated?: Integer;
    linesOfCodeAccepted?: Integer;
    charsOfCodeGenerated?: Integer;
    charsOfCodeAccepted?: Integer;
  }
// ...
  export interface TransformEvent {
    jobId: TransformationJobId;
    timestamp?: Timestamp;
    ideCategory?: IdeCategory;
    programmingLanguage?: ProgrammingLanguage;
    linesOfCodeChanged?: Integer;
    charsOfCodeChanged?: Integer;
    linesOfCodeSubmitted?: Integer;
  }
// ...
  export type UUID = string;
// ...
  export interface UserContext {
    ideCategory: IdeCategory;
    operatingSystem: OperatingSystem;
    product: UserContextProductString;
    clientId?: UUID;
    ideVersion?: String;
  }
  export type UserContextProductString = string;
// ...
  export interface UserModificationEvent {
    sessionId: UUID;
    requestId: UUID;
    programmingLanguage: ProgrammingLanguage;
    modificationPercentage: Double;
    customizationArn?: CustomizationArn;
    timestamp: Timestamp;
    acceptedCharacterCount: PrimitiveInteger;
    unmodifiedAcceptedCharacterCount: PrimitiveInteger;
  }
  export interface UserSettings {
    hasConsentedToCrossRegionCalls?: Boolean;
  }
  export interface UserTriggerDecisionEvent {
    sessionId: UUID;
    requestId: UUID;
    customizationArn?: CustomizationArn;
    programmingLanguage: ProgrammingLanguage;
    completionType: CompletionType;
    suggestionState: SuggestionState;
    recommendationLatencyMilliseconds: Double;
    timestamp: Timestamp;
    triggerToResponseLatencyMilliseconds?: Double;
    suggestionReferenceCount?: PrimitiveInteger;
    generatedLine?: PrimitiveInteger;
    numberOfRecommendations?: PrimitiveInteger;
    perceivedLatencyMilliseconds?: Double;
    acceptedCharacterCount?: PrimitiveInteger;
    addedIdeDiagnostics?: IdeDiagnosticList;
    removedIdeDiagnostics?: IdeDiagnosticList;
  }